refactor(src): restructure backend into src/chatty_commander with shims; update packaging
