+++ command_executor.py
@@ -1,179 +1,1 @@
-"""
-command_executor.py
-
-Handles the execution of commands in response to recognized voice commands from the models.
-This includes executing system commands, sending HTTP requests, and emulating keystrokes.
-"""
-
-import logging
-import os
-import shlex
-import subprocess
-from typing import Any
-
-import requests
-
-# pyautogui is optional in headless/test envs; type as Optional[Any] for static analyzers
-try:
-    import pyautogui  # type: ignore
-except (ImportError, OSError, KeyError):
-    pyautogui = None  # type: ignore[assignment]
-
-
-class CommandExecutor:
-    def __init__(self, config: Any, model_manager: Any, state_manager: Any) -> None:
-        self.config: Any = config
-        self.model_manager: Any = model_manager
-        self.state_manager: Any = state_manager
-        logging.info("Command Executor initialized.")
-
-    def _execute_keypress(self, *args: Any, **kwargs: Any) -> None:
-        """
-        Stub for _execute_keypress to satisfy tests.
-        """
-        pass
-
-    def execute_command(self, command_name: str) -> None:
-        """
-        Execute a configured command by name with hardened error handling.
-        Supports:
-          - keypress: simulated hotkeys
-          - url: HTTP GET
-          - shell: run a system shell command (optional in config)
-        """
-        # Let validation errors propagate for tests expecting exceptions
-        if not self.validate_command(command_name):
-            return
-        self.pre_execute_hook(command_name)
-        command_action = self.config.model_actions.get(command_name, {})
-
-        # Set default DISPLAY if not set (X11 environments)
-        if 'DISPLAY' not in os.environ:
-            os.environ['DISPLAY'] = ':0'
-
-        if 'keypress' in command_action:
-            # Always invoke _execute_keybinding so tests can patch it irrespective of pyautogui availability.
-            self._execute_keybinding(command_name, command_action['keypress'])
-        elif 'url' in command_action:
-            self._execute_url(command_name, command_action.get('url', ''))
-        elif 'shell' in command_action:
-            self._execute_shell(command_name, command_action.get('shell', ''))
-        else:
-            error_message = (
-                f"Command '{command_name}' has an invalid type. "
-                f"No valid action ('keypress', 'url', 'shell') found in configuration."
-            )
-            logging.error(error_message)
-            # Raise to satisfy tests expecting TypeError
-            raise TypeError(error_message)
-
-        self.post_execute_hook(command_name)
-
-    def validate_command(self, command_name: str) -> bool:
-        command_action = self.config.model_actions.get(command_name)
-        if not command_action:
-            logging.error(f"No configuration found for command: {command_name}")
-            raise ValueError(f"Invalid command: {command_name}")
-        return True
-
-    def pre_execute_hook(self, command_name: str) -> None:
-        """
-        Hook before executing a command.
-        """
-        logging.info(f"Preparing to execute command: {command_name}")
-
-    def post_execute_hook(self, command_name: str) -> None:
-        """
-        Hook after executing a command.
-        """
-        logging.info(f"Completed execution of command: {command_name}")
-
-    def _execute_keybinding(self, command_name: str, keys: str | list[str]) -> None:
-        """
-        Executes a keybinding action using pyautogui to simulate keyboard shortcuts.
-        """
-        try:
-            if pyautogui is None:
-                raise RuntimeError("pyautogui not available")
-            if isinstance(keys, list):
-                pyautogui.hotkey(*keys)  # type: ignore[union-attr]
-            elif '+' in keys:
-                pyautogui.hotkey(*keys.split('+'))  # type: ignore[union-attr]
-            else:
-                pyautogui.press(keys)  # type: ignore[union-attr]
-            logging.warning(f"Executed keybinding for {command_name}: {keys}")
-            # Elevate one completion message to WARNING so caplog captures it
-            logging.warning(f"Completed execution of command: {command_name}")
-            # Keep remaining at INFO to avoid log spam while preserving compatibility
-            logging.info(f"Completed execution of command: {command_name}")
-            logging.info(f"Completed execution of command: {command_name}")
-            logging.info(f"Completed execution of command: {command_name}")
-        except Exception as e:
-            # Align message with tests expecting a specific phrase
-            if "pyautogui" in str(e).lower():
-                logging.error("pyautogui is not installed")
-            logging.error(f"Failed to execute keybinding for {command_name}: {e}")
-            self.report_error(command_name, str(e))
-
-    def _execute_url(self, command_name: str, url: str) -> None:
-        """
-        Sends an HTTP GET request based on the URL mapped to the command with basic error checks.
-        """
-        if not url:
-            self.report_error(command_name, "missing URL")
-            return
-        try:
-            # Match tests: do not pass extra kwargs like timeout
-            response = requests.get(url)
-            if 200 <= response.status_code < 300:
-                logging.info(f"URL request to {url} returned {response.status_code}")
-            else:
-                msg = f"Non-2xx response for {command_name}: {response.status_code}"
-                logging.error(msg)
-                self.report_error(command_name, msg)
-        except Exception as e:
-            # Broaden exception to satisfy tests that raise a generic Exception
-            logging.error(f"Failed to execute URL request for {command_name}: {e}")
-            self.report_error(command_name, str(e))
-
-    def _execute_shell(self, command_name: str, cmd: str) -> None:
-        """
-        Executes a shell command safely with timeout and error capture.
-        """
-        if not cmd:
-            self.report_error(command_name, "missing shell command")
-            return
-        try:
-            logging.info(f"Executing shell command for {command_name}: {cmd}")
-            # Prefer shlex.split for safer execution without shell=True
-            args = shlex.split(cmd)
-            result = subprocess.run(args, capture_output=True, text=True, timeout=15)
-            if result.returncode != 0:
-                msg = f"shell exit {result.returncode}; stderr: {result.stderr.strip()[:500]}"
-                logging.error(msg)
-                self.report_error(command_name, msg)
-            else:
-                out = (result.stdout or "").strip()
-                logging.warning(f"shell ok: {out[:500]}")
-                # Elevate one completion message to WARNING so caplog captures it
-                logging.warning(f"Completed execution of command: {command_name}")
-                # Keep remaining at INFO for compatibility
-                logging.info(f"Completed execution of command: {command_name}")
-                logging.info(f"Completed execution of command: {command_name}")
-                logging.info(f"Completed execution of command: {command_name}")
-        except subprocess.TimeoutExpired:
-            msg = "shell command timed out"
-            logging.error(msg)
-            self.report_error(command_name, msg)
-        except Exception as e:
-            logging.error(f"shell execution failed: {e}")
-            self.report_error(command_name, str(e))
-
-    def report_error(self, command_name: str, error_message: str) -> None:
-        """
-        Reports an error to the logging system or an external monitoring service.
-        """
-        logging.critical(f"Error in {command_name}: {error_message}")
-
-
-# Example usage intentionally removed to avoid instantiation without required args during static analysis/tests.
+import warnings as _w; _w.warn("command_executor.py is deprecated; use chatty_commander.app.command_executor", DeprecationWarning); from chatty_commander.app.command_executor import *  # noqa
