+++ src/chatty_commander/app/state_manager.py
